+++ tests/DependencyInjection/Compiler/CmfRouterAutoRegisterTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Sonata\PageBundle\Tests\DependencyInjection;
+namespace Sonata\PageBundle\Tests\DependencyInjection\Compiler;
 
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractCompilerPassTestCase;
 use Sonata\PageBundle\DependencyInjection\Compiler\CmfRouterCompilerPass;
