+++ tests/DependencyInjection/SonataPageExtensionTest.php
@@ -16,7 +16,6 @@ namespace Sonata\PageBundle\Tests\DependencyInjection;
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractExtensionTestCase;
 use Sonata\PageBundle\DependencyInjection\SonataPageExtension;
 use Symfony\Bundle\TwigBundle\DependencyInjection\TwigExtension;
-use Twig\Extra\String\StringExtension;
 
 /**
  * @author Rémi Marseille <marseille@ekino.com>
@@ -97,15 +96,6 @@ class SonataPageExtensionTest extends AbstractExtensionTestCase
             '@SonataForm/Form/datepicker.html.twig',
             $this->container->getParameter('twig.form.resources')
         );
-    }
-
-    public function testLoadTwigStringExtension(): void
-    {
-        $this->container->setParameter('kernel.bundles', []);
-
-        $this->load();
-
-        $this->assertContainerBuilderHasServiceDefinitionWithTag(StringExtension::class, 'twig.extension');
     }
 
     protected function getContainerExtensions(): array
